fix(modList): restrict paths that can be specified
