+++ src/lib/apmPath.ts
@@ -1,0 +1,30 @@
+import path from 'path';
+
+/**
+ * Determine the parent-child relationship of a path.
+ *
+ * @param {string} parent - Path expected to be a parent folder.
+ * @param {string} child - Paths expected to be a child entry.
+ * @returns {boolean} - Boolean value
+ */
+const isParent = (parent: string, child: string) => {
+  const relative = path.relative(parent, child);
+  return relative && relative !== '' && !relative.startsWith('..');
+};
+
+/**
+ * Determine if two paths have a parent-child relationship.
+ *
+ * @param {string} pathA - A path
+ * @param {string} pathB - A path
+ * @returns {boolean} - Boolean value
+ */
+const pathRelated = (pathA: string, pathB: string) => {
+  return isParent(pathA, pathB) || isParent(pathB, pathA);
+};
+
+const apmPath = {
+  isParent,
+  pathRelated,
+};
+export default apmPath;
